Add keychain providers along with shell provider
Keychain providers are designed to provide a bridge between this plugin
and user's secret storage. It may be system keychain/keyring service or
some password manager; we don't care much as long as it can be invoked
and securely return a secret.

Shell provider is the most basic provider that does nothing securily.
However, this is an ultimate bridge that can be used with any custom
script or CLI program. Later on, we will implement system keychain
provider.