Check user provided data for illegal characters
Some of existing auth providers use passed data as HTTP header names
and/or values. We've got to validate these data because HTTP sets
restriction on what characters may or may not be used in HTTP headers.

This does not affect all auth providers as some of them may use base64
digest instead.