Use Ruff instead of flake8/black
Ruff is a new toy in town. It's to install, it comes with many batteries
included, it has a pretty good integration into text editors and it
boosts my productivity. So this patch switches to Ruff from flake8 and
black.

In addition to that switch, this patch also drops the pre-commit
configuration since it actually never worked for me.