Turn most Ruff lint rules on
Default Ruff configuration tries to be as close as possible to Flake8,
so that the former can be considered a drop-in replacement of the
latter. Ruff, however, comes with many batteries included, and there are
many useful rules out there. This patch enables most of them, leaving
our the one the code base is not ready for for future.