Decrease the size of the testing matrix
This plugin claims to support 5 Python versions: 3.8 to 3.12, and 3
operating systems: Linux, macOS and Windows. By testing all possible
combinations, it requires running 15 test jobs in parallel.

Unfortunately, Windows & macOS runners aren't fast to be found, and
sometimes one has to wait a significant amount of time for a pull
request to become "green". This is wasteful for me, as a sole
contributor, and GitHub, as a company that provides compute resources
for free.

In practice there's almost no benefits of testing all python versions on
Windows and macOS platforms as long as we test them elsewhere (Linux).
In most cases, either the python version is not supported regardless of
the platform it runs on, or the platform is not supported regardless of
the python version you run on it.

This patch decreases the testing matrix size by running tests on Windows
and macOS runners just once, using the latest available python version.