Modernize tox-poetry integration
I have never been a fan of poetry, but since this project is already
using it, let's keep it this way. What we have to do though is to make
sure we use the up-to-date poetry configuration and best practices.

This patch moves dependency definitions out from tox.ini to
pyproject.toml (poetry groups). This should help us to maintain
dependencies in one place only.