password-store: check returned secret in the test
The test we have today doesn't check retrieved secret except for its
length. Even though in most cases testing the length most likely will be
enough, it's not robust enough. This patch makes the test better by
creating a known secret in password-store, retrieving it and then
testing that the secret is expected.