Use plugin via '-A store', deprecate '-A creds'
The `-A` (`--auth-type`) parameter stands for authentication type, and
is essentially the ID of the authentication plugin to invoke. This patch
deprecates both `-A credential-store` and `-A creds` usages in favor of
newly added `-A store`.

I believe that `store` is both convenient and transparent, so we better
deprecate other aliases and remove them in the future in order to
prevent polluting plugins namespace.