Remove tests/conftest.py w/ a hack
The conftest.py today contains a hack that creates a session scoped
temporary directory, and set the HTTPIE_CONFIG_DIR environment variable
to that temporary directory. That previously has been done to make sure
I can run tests with configuration store elsewhere. Unfortunately, this
is a bit inconvenient because (1) we have to make sure there's no
imports to httpie package in global scope because otherwise the patch we
have won't work, (2) the creation and cleaning phase is scatted across
the code base.

This patch removes this hack in favor of (1) explicitly passing the
temporary directory to executing httpie environment context, and (2)
patching required attribute so the plugin can use the up to date version
whenever it runs.