Auth store format v2
First version of the auth store lacked two major features: (1) it was
impossible to share secrets between auth bindings, and (2) it was
impossible to define an auth binding to many resources (URLs)
simultaneously. In addition to that, it contained a bunch of nested
blocks which were hard to write and mantain.

New, second version of the auth store is designed to address those
concerns by (1) separating secrets from auth bindings, (2) reducing
nesting level, and (3) supporting multiple resources (URLs) per auth
binding.

    {
        "bindings": [
            {
                "auth_type": "bearer",
                "auth": "$GITHUB_TOKEN",
                "resources": ["https://api.github.com/"]
            }
        ],
        "secrets": {
            "GITHUB_TOKEN": {
                "provider": "system",
                "service": "github.com",
                "username": "ikalnytskyi"
            }
        }
    }

This patch also delivers a significant internal refactoring. The most
notable change is that "types" are enforced internally instead of
dealing with dictionaries around. I don't think performance is going to
be an issue here. :)